feat: ResponsiveImage refactoring
