+++ apps/www/src/components/PhotoCard.tsx
@@ -1,9 +1,6 @@
-'use client'
-
 import { FC } from 'react'
-import NextImage from 'next/image'
-import { Card, CardFooter } from '@nextui-org/card'
 
+import { ResponsiveImage } from 'src/components/ResponsiveImage'
 import { getT } from 'src/utils/getT'
 
 import avatarImg1200 from 'public/avatar.jpg'
@@ -19,18 +16,13 @@ const sizes = images
 
 export const PhotoCard: FC<WithLangProp> = ({ lang }) => {
   const t = getT(lang)
-
-  return (
-    <Card as="figure" className="block relative" isFooterBlurred>
-      <NextImage sizes={sizes} src={avatarImg800} alt={t.fullName} priority={false} />
-      <CardFooter
-        as="figcaption"
-        className="before:bg-white/10 py-1 block absolute before:rounded-xl rounded-lg text-tiny text-white/80 bottom-2 right-2 max-w-fit ml-1 z-10"
-      >
-        {t.fullName}
-        <br />
-        📷 Alina Delyne
-      </CardFooter>
-    </Card>
+  const caption = (
+    <>
+      {t.fullName}
+      <br />
+      📷 Alina Delyne
+    </>
   )
+
+  return <ResponsiveImage alt={t.fullName} caption={caption} sizes={sizes} src={avatarImg800} />
 }
